Removed interface MailServer
